src/live_trading/tests/test_trade_module/requirements.txt
requests
websocket-client
slack_sdk